Fixed Search Bug
This style tag was causing issues for the search bar. I have moved it  outside of the table element so it should not affect searching anymore. Please note that any html elements inside the table element is inculded as part of the search, so refrain from adding anything if you do not want it included in the search.